+++ aws-android-sdk-core/src/main/java/com/amazonaws/logging/LogFactory.java
@@ -23,6 +23,11 @@ import java.util.HashMap;
  * for backwards compatibility.
  */
 public class LogFactory {
+
+    /**
+     * NOTE : Any changes to rename this class should ensure that this log tag is no longer than 23.
+     * Log tag longer than 23 will cause it to break on Android API level <= 23.
+     */
     private static final String TAG = LogFactory.class.getSimpleName();
     private static final String APACHE_COMMONS_LOGGING_LOGFACTORY = "org.apache.commons.logging.LogFactory";
 
@@ -35,29 +40,31 @@ public class LogFactory {
      * @return logger
      */
     public static synchronized Log getLog(Class clazz) {
-        return getLog(clazz.getSimpleName());
+        return getLog(getTruncatedLogTag(clazz.getSimpleName()));
     }
 
     /**
      * Get the logger for the string tag
-     * @param string the string tag
+     * @param logTag the string tag
      *
      * @return logger
      */
-    public static synchronized Log getLog(final String string) {
-        Log log = logMap.get(string);
+    public static synchronized Log getLog(String logTag) {
+        logTag = getTruncatedLogTag(logTag);
+
+        Log log = logMap.get(logTag);
         if (log == null) {
             if (checkApacheCommonsLoggingExists()) {
                 try {
-                    log = new ApacheCommonsLogging(string);
-                    logMap.put(string, log);
+                    log = new ApacheCommonsLogging(logTag);
+                    logMap.put(logTag, log);
                 } catch (Exception e) {
                     android.util.Log.w(TAG, "Could not create log from " + APACHE_COMMONS_LOGGING_LOGFACTORY, e);
                 }
             }
             if (log == null) {
-                log = new AndroidLog(string);
-                logMap.put(string, log);
+                log = new AndroidLog(logTag);
+                logMap.put(logTag, log);
             }
         }
         return log;
@@ -73,5 +80,25 @@ public class LogFactory {
             android.util.Log.e(TAG, ex.getMessage());
             return false;
         }
+    }
+
+    /**
+     * Truncate log tag to be within 23 characters in length as required by Android on certain API levels.
+     *
+     * @param logTag Log tag to be truncated
+     * @return truncated log tag
+     */
+    private static String getTruncatedLogTag(String logTag) {
+        if (logTag.length() > 23) {
+            if (checkApacheCommonsLoggingExists()) {
+                Log log = new ApacheCommonsLogging(TAG);
+                log.warn("Truncating log tag length as it exceed 23, the limit imposed by Android on certain API Levels");
+            } else {
+                android.util.Log.w(TAG, "Truncating log tag length as it exceed 23, the limit imposed by Android on certain API Levels");
+            }
+            logTag = logTag.substring(0, 23);
+        }
+
+        return logTag;
     }
 }